utils: iio_info: Explicitely cast error codes to int
iio_strerror() expects an "int" value for the error code. Visual Studio
will complain if we are passing a ssize_t, as it is a downcast. However,
we know that error codes all fit in "int", so the downcast is safe.

+++ utils/iio_info.c
@@ -90,7 +90,7 @@ static void print_attr(const struct iio_attr *attr,
 	if (!value) {
 		ret = iio_attr_read_raw(attr, buf, sizeof(buf) - 1);
 		if (ret < 0)
-			iio_strerror(ret, buf, sizeof(buf));
+			iio_strerror((int)ret, buf, sizeof(buf));
 		value = buf;
 	}
